# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple weather server.",
        "title": "Swagger Example API of the weather server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "paths": {
        "/detectors": {
            "get": {
                "description": "get all detectors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List detectors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Detector"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create detector",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create detector",
                "parameters": [
                    {
                        "description": "Detector",
                        "name": "detector",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Detector"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Detector"
                        }
                    }
                }
            }
        },
        "/detectors/{id}": {
            "get": {
                "description": "get detector by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get detector by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of detector",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Detector"
                        }
                    }
                }
            },
            "post": {
                "description": "update detector",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update detector",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of detector",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Detector",
                        "name": "detector",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Detector"
                        }
                    }
                ]
            },
            "delete": {
                "description": "delete detector",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete detector",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of detector",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        }
    },
    "definitions": {
        "model.Detector": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                }
            }
        }
    }
}